[scala] ExecutorManager refactor
