+++ scala-package/core/src/main/scala/ml/dmlc/mxnet/Executor.scala
@@ -1,8 +1,6 @@
 package ml.dmlc.mxnet
 
 import ml.dmlc.mxnet.Base._
-import org.slf4j.{Logger, LoggerFactory}
-
 import scala.collection.mutable.ArrayBuffer
 
 object Executor {
@@ -11,94 +9,6 @@ object Executor {
                              ndarrays: Seq[NDArray]): Map[String, NDArray] = {
     require(names.toSet.size == names.length, "Duplicate names detected")
     (names zip ndarrays).toMap
-  }
-
-  /**
-   * Get input slice from the input shape.
-   * @param batchSize The number of samples in a mini-batch.
-   * @param workLoadList The list of work load for different devices, in the same order as ctx
-   * @return The split slices to get a specific slice.
-   * @throws IllegalArgumentException
-   *         If there are two many splits such that some slice can be empty.
-   */
-  private[mxnet] def splitInputSlice(batchSize: Int,
-                                     workLoadList: Seq[Float]): Array[(Int, Int)] = {
-    val totalWorkLoad = workLoadList.sum
-    val batchNumList = workLoadList.map(workLoad =>
-      math.round(workLoad * batchSize / totalWorkLoad)).toArray
-    val batchNumSum = batchNumList.sum
-    if (batchNumSum < batchSize) {
-      batchNumList(batchNumList.length-1) += batchSize - batchNumSum
-    }
-
-    val slices = ArrayBuffer.empty[(Int, Int)]
-    var end = 0
-    batchNumList.foreach(batchNum => {
-      val begin = math.min(end, batchSize)
-      end = math.min(begin + batchNum, batchSize)
-      require(begin < end, "Too many slices such that some splits are empty")
-      slices.append((begin, end))
-    })
-    slices.toArray
-  }
-
-  /**
-   * Check the argument names of symbol.
-   * This function checks the duplication of arguments in Symbol.
-   * The check is done for feedforward net for now.
-   * @param symbol The network configuration
-   */
-  private[mxnet] def checkArguments(symbol: Symbol): Unit = {
-    val argNames = symbol.listArguments()
-    require(argNames.toSet.size == argNames.length,
-      "Find duplicated argument name," +
-      "please make the weight name non-duplicated(using name arguments)," +
-      s"arguments are $argNames")
-
-    val auxNames = symbol.listAuxiliaryStates()
-    require(auxNames.toSet.size == auxNames.length,
-      "Find duplicated auxiliary param name," +
-      "please make the weight name non-duplicated(using name arguments)," +
-      s"arguments are $auxNames")
-  }
-
-  // Load a list of arrays into a list of arrays
-  private[mxnet] def loadGeneral(data: Seq[NDArray], targets: Seq[NDArray]): Unit = {
-    (data zip targets).foreach { case (dSrc, dTarget) =>
-      dSrc.copyTo(dTarget)
-    }
-  }
-
-  // Load a list of arrays into a list of arrays specified by slices
-  private[mxnet] def loadGeneralMulti(data: Seq[NDArray],
-                                      targets: Seq[Array[(Int, Int, NDArray)]]): Unit = {
-    for ((src, dTargets) <- data zip targets) {
-      for ((start, end, dst) <- dTargets) {
-        val sliced = src.slice(start, end)
-        sliced.copyTo(dst)
-        sliced.dispose()
-      }
-    }
-  }
-
-  // Load data into sliced arrays
-  private[mxnet] def loadDataMulti(batch: DataBatch,
-                                   targets: Seq[Array[(Int, Int, NDArray)]]): Unit = {
-    loadGeneralMulti(batch.data, targets)
-  }
-
-  private[mxnet] def loadData(batch: DataBatch, targets: Seq[NDArray]): Unit = {
-    loadGeneral(batch.data, targets)
-  }
-
-  // Load label into sliced arrays
-  private[mxnet] def loadLabelMulti(batch: DataBatch,
-                                    targets: Seq[Array[(Int, Int, NDArray)]]): Unit = {
-    loadGeneralMulti(batch.label, targets)
-  }
-
-  private[mxnet] def loadLabel(batch: DataBatch, targets: Seq[NDArray]): Unit = {
-    loadGeneral(batch.label, targets)
   }
 }
 
@@ -124,6 +34,7 @@ class Executor private[mxnet](private[mxnet] val handle: ExecutorHandle,
   private[mxnet] var auxArrays: Array[NDArray] = null
   val outputs: Array[NDArray] = getOutputs
   protected var _argDict: Map[String, NDArray] = null
+  protected var _gradDict: Map[String, NDArray] = null
   protected var _auxDict: Map[String, NDArray] = null
   protected var monitorCallback: MXMonitorCallback = null
   private[mxnet] var _ctx: Context = null
@@ -196,7 +107,7 @@ class Executor private[mxnet](private[mxnet] val handle: ExecutorHandle,
     }
 
     var newAuxDict = Map[String, NDArray]()
-    val zip3 = (this.symbol.listAuxiliaryStates, auxShapes, this.auxArrays).zipped
+    val zip3 = (this.symbol.listAuxiliaryStates(), auxShapes, this.auxArrays).zipped
     zip3.foreach { case (name, newShape, arr) =>
       if (partialShaping || newShape.equals(arr.shape)) {
         if (newShape.product > arr.shape.product) {
@@ -306,6 +217,18 @@ class Executor private[mxnet](private[mxnet] val handle: ExecutorHandle,
   }
 
   /**
+   * Get dictionary representation of gradient arrays.
+   * @return The dictionary that maps name of arguments to gradient arrays.
+   * @throws IllegalArgumentException if there are duplicated names in the grads.
+   */
+  def gradDict: Map[String, NDArray] = {
+    if (_gradDict == null) {
+      _gradDict = Executor.getDict(symbol.listArguments(), gradArrays)
+    }
+    _gradDict
+  }
+
+  /**
    * Get dictionary representation of auxiliary states arrays.
    * @return The dictionary that maps name of auxiliary states to NDArrays.
    * @throws IllegalArgumentException if there are duplicated names in the auxiliary states.
@@ -368,151 +291,3 @@ class Executor private[mxnet](private[mxnet] val handle: ExecutorHandle,
   }
 }
 // scalastyle:on finalize
-
-/**
- * Helper class to manage multiple executors for data parallelism.
- * @author Yizhi Liu
- * @param symbol output symbol
- * @param ctx devices to run on
- * @param paramNames Name of all trainable parameters of the network.
- * @param argNames Name of all arguments of the network.
- * @param auxNames Name of all auxiliary states of the network.
- * @param trainData Training data iterator.
- * @param workLoadList The list of work load for different devices, in the same order as ctx
- * @param logger When not specified, default logger will be used.
- */
-class DataParallelExecutorManager(symbol: Symbol,
-                                  ctx: Array[Context],
-                                  paramNames: Seq[String],
-                                  argNames: Seq[String],
-                                  private val auxNames: Seq[String],
-                                  trainData: DataIter,
-                                  private var workLoadList: Seq[Float] = null,
-                                  logger: Logger = DataParallelExecutorManager.logger) {
-  // preparation
-  private val numDevice = ctx.length
-  logger.info(s"Start training with [${ctx.mkString(",")}]")
-
-  // make sure the architecture is valid
-  Executor.checkArguments(symbol)
-
-  if (workLoadList == null) {
-    workLoadList = Seq.fill(numDevice)(1f)
-  }
-  require(workLoadList.size == numDevice, "Invalid settings for work load.")
-
-  private val slices = Executor.splitInputSlice(trainData.batchSize, workLoadList)
-
-  private val trainExecs =
-    ctx.zipWithIndex.map { case (context, i) =>
-      val dataShapes =
-        (trainData.provideData ++ trainData.provideLabel).map { case (name: String, shape: Shape) =>
-          (name, Shape(slices(i)._2 - slices(i)._1) ++ shape.drop(1))
-        }
-      symbol.simpleBind(context, "write", shapeDict = dataShapes)
-    }
-
-  // data structure
-  private val dataNames = trainData.provideData.map(_._1).toArray
-  private val labelNames = trainData.provideLabel.map(_._1).toArray
-
-  private val dataArrays =
-    dataNames.map { name =>
-      trainExecs.zipWithIndex.map { case (exec, i) =>
-        val slice = slices(i)
-        (slice._1, slice._2, exec.argDict(name))
-      }
-    }
-  private val labelArrays =
-    labelNames.map { name =>
-      trainExecs.zipWithIndex.map { case (exec, i) =>
-        val slice = slices(i)
-        (slice._1, slice._2, exec.argDict(name))
-      }
-    }
-
-  private val paramIdx = (0 until argNames.length).filter { i =>
-    paramNames.contains(argNames(i))
-  }
-  private[mxnet] val _paramNames = paramIdx.map(argNames(_))
-  private[mxnet] val paramArrays = paramIdx.map { i =>
-    trainExecs.map(_.argArrays(i))
-  }.toArray
-  private[mxnet] val gradArrays = paramIdx.map { i =>
-    trainExecs.map(_.gradArrays(i))
-  }.toArray
-
-  private val auxArrays = (0 until auxNames.length).map { i =>
-    trainExecs.map(_.auxArrays(i))
-  }.toArray
-  private val batchSize = trainData.batchSize
-  private val outputShapes: Array[Shape] = trainExecs(0).outputs.map { x: NDArray =>
-      Shape(batchSize) ++ x.shape.drop(1)
-    }
-  private[mxnet] val cpuOutputArrays = outputShapes.map(NDArray.zeros(_))
-
-  /**
-   * Release the related executors.
-   * The object shall never be used after it is disposed.
-   */
-  def dispose(): Unit = {
-    trainExecs.foreach(_.dispose())
-  }
-
-  // Install monitor on all executors
-  def installMonitor(monitor: Monitor): Unit = {
-    trainExecs.foreach(monitor.install)
-  }
-
-  /**
-   * Set parameter and aux values
-   * @param argParams source parameter arrays
-   * @param auxParams source aux arrays
-   */
-  def setParams(argParams: Map[String, NDArray], auxParams: Map[String, NDArray]): Unit = {
-    trainExecs.foreach(_.copyParamsFrom(argParams, auxParams))
-  }
-
-  /**
-   * Copy data from each executor to `arg_params` and `aux_params`
-   * @param argParams target parameter arrays
-   * @param auxParams target aux arrays
-   * @note This function will inplace update the NDArrays in arg_params and aux_params.
-   */
-  def copyTo(argParams: Map[String, NDArray], auxParams: Map[String, NDArray]): Unit = {
-    for ((name, block) <- _paramNames zip paramArrays) {
-      val weight = block.map(_.copyTo(Context.cpu())).reduce(_ + _) / block.length
-      weight.copyTo(argParams(name))
-    }
-    for ((name, block) <- auxNames zip auxArrays) {
-      val weight = block.map(_.copyTo(Context.cpu())).reduce(_ + _) / block.length
-      weight.copyTo(auxParams(name))
-    }
-  }
-
-  // load data and labels into arrays
-  def loadDataBatch(dataBatch: DataBatch): Unit = {
-    Executor.loadDataMulti(dataBatch, dataArrays)
-    Executor.loadLabelMulti(dataBatch, labelArrays)
-  }
-
-  // Perform a forward pass on each executor
-  def forward(isTrain: Boolean = false): Unit = {
-    for ((texec, islice) <- trainExecs zip slices) {
-      texec.forward(isTrain)
-      for ((cpuOut, devOut) <- cpuOutputArrays zip texec.outputs) {
-        devOut.copyTo(cpuOut.slice(islice))
-      }
-    }
-  }
-
-  // Perform a backward pass on each executor
-  def backward(): Unit = {
-    trainExecs.foreach(_.backward())
-  }
-}
-
-object DataParallelExecutorManager {
-  private val logger = LoggerFactory.getLogger(classOf[Model])
-}
-
